Resolve term-length / role for IETF chair / IESG slot
In Term Lengths there are 2 options. We need to pick one after the merge and remove the other.

I also removed "An exception is if a Director role is occupied by the IETF Chair ex officio, since that person’s service is governed instead by the term lengths established in {{RFC7437}} (BCP10), Section 3.4." from Term Limits as it seemed duplicative.