Lear change 2
As to this:

   All requests
   for review shall be posted publicly, 

This is necessary but not sufficient.  You need to be clear as to what this means, lest it be claimed that they did so and you didn’t follow process.  My suggestion is to replace it with something like:
   All requests
   for review shall be emailed to llc@ietf.org who SHALL in turn publish them.


Actual change was:
ll requests for
review shall be sent to the IETF LLC, such as via email to llc@ietf.org. 
The IETF LLC shall in turn publish those publicly, and is expected to
respond publicly to these requests within a reasonable period,
typically within 90 days.