IESG review 3.8
removal of 3.8 as superfluous

Comment:
This section seems superfluous. This document is intended to be published as a consensus IETF BCP, so the text here is trivially true, just as with all consensus-based RFCs.

## Community Consensus and Grant of Authority

The IETF is a consensus-based group, and authority to act on behalf of
the community requires a high degree of consensus and the continued
consent of the community.  After a careful process of deliberation, a
broad-based community consensus emerged to house the administration of the 
IETF within the IETF LLC as a disregarded entity of the Internet Society. 
This document reflects that consensus.